+++ projects/core/src/remult3/remult3.ts
@@ -818,7 +818,6 @@ export type ClassFieldDecorator<entityType, valueType> = (
 
 //p1 - `preprocessApiFindOptions`,  naming - allow find options preprocessor for api calls, to use for authorization
 //p1 - admin url1 - pocketbase use /_ for the admin //{allowed?:Allowed,url?:string}
-//p1 - Validate for NAN
 //p1 - Make async hooks gracefully handle non async hooks environment
 //p1 - chain validators with options